+++ core/blockchain.ts/CHANGELOG.md
@@ -8,6 +8,6 @@ The changelog format is based on [Keep a Changelog](https://keepachangelog.com/e
 #### Added
 
 - feat(api): add Reader and Writer classes to handle buffering
-- feat(api): add a Template engine based on Writer
+- feat(api): add a Template engine based on Writer class
 - feat(api): add abstract classes Block, Transaction and Wallet
 - feat(api): add MerkleTree class